backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-jose[cryptography]>=3.3.0
argon2-cffi>=23.1.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
aiosqlite>=0.20.0
python-multipart>=0.0.18
